Basis build - Python - Flask/sqlite3
This is an incomplete basis for our project, I reused the code from a recent flask project I was asked to do, I have commented out the code that was propritary to that project but did not delete that code so that it may be used for reference. Ideally we will move away from this stack when we upscale.