Recipe-Pantry Comparison
Implemented a method to use list.html to render ranked recipe choices based on matching ingredients. Currently the buttons display the number of ingredients met and needed. I intend to make the list.html show recipe suggestions in a page section view. This allow to show photos (if implemented) and info about the recipe (brief description, numSteps/Ingredients, normalize the numSteps/Ingredients to map to a simple difficulty range) or reviews.